chore(accordion): remove legacy context after compat shim
